# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ride-Hailing Platform API
  version: 1.0.0
  description: |
    REST API for riders, drivers, operations staff and admins. Design: [HLD §10](architecture.md#10-api-design),
    [LLD §13](low-level-design.md#13-rest-api).

    - **Authentication**: `Authorization: Bearer <access token>` from `/v1/auth/token` (JWT, 15 min). Sign-in and
      webhook endpoints are public; webhooks are signed.
    - **Idempotency**: commands that change a ride, offer, driver status, payment or rating require an
      `Idempotency-Key` (1–255 visible ASCII characters). A retry with the same key and request returns the original
      response with `Idempotent-Replayed: true`; the same key with a different request gets
      `422 IDEMPOTENCY_KEY_REUSED`; a key still executing gets `409 IDEMPOTENCY_KEY_IN_PROGRESS` with `Retry-After`.
      Keys are kept for 24 hours.
    - **Format**: JSON in snake_case. Money is `{"amount_paise": 25900, "currency": "INR"}`. Timestamps are
      RFC 3339 in UTC. Fields without a value are omitted. Clients must ignore unknown fields.
    - **Errors**: RFC 9457 problem details (`application/problem+json`) with a stable `code` and the `request_id`.
      Any operation may also answer `500 INTERNAL_ERROR` or `503 SERVICE_UNAVAILABLE` (with `Retry-After`); retry
      commands with the same `Idempotency-Key` and backoff.
    - **Visibility**: a resource the caller may not see answers `404`, not `403`.
...

...
servers:
...
security:
  - bearer: []
...
paths:
...
  /v1/drivers/me:
    get:
      tags: [Drivers]
      operationId: getDriver
      summary: The signed-in driver's profile, vehicles and status
      responses:
        '200':
          description: Driver.
          headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Driver'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '404': {$ref: '#/components/responses/NotFound'}
  /v1/drivers/me/online:
    post:
      tags: [Drivers]
      operationId: goOnline
      summary: Go online with a vehicle
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GoOnlineRequest'}
      responses:
        '200': {$ref: '#/components/responses/DriverStatus'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/drivers/me/offline:
    post:
      tags: [Drivers]
      operationId: goOffline
      summary: Go offline
      description: Refused during a ride. A pending offer is declined first.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200': {$ref: '#/components/responses/DriverStatus'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '409': {$ref: '#/components/responses/Conflict'}
  /v1/drivers/me/location:
    post:
      tags: [Drivers]
      operationId: sendLocations
      summary: Location updates over HTTP
      description: |
        V1: every live update (one per request, every 4 s). From V2 live updates use the WebSocket, and this endpoint
        carries offline replays of up to 100 updates marked `replay: true`. Updates not newer than the last applied
        sequence number are ignored. Limited to 1 request per second per driver.
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LocationBatch'}
            example:
              updates:
                - {seq: 1842, lat: 12.97571, lon: 77.60502, accuracy_m: 8.5, heading_deg: 91, speed_mps: 7.2,
                   device_time: '2026-10-02T08:15:04.120Z', replay: false}
      responses:
        '200':
          description: Outcome counts.
          headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/LocationBatchResult'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/TooManyRequests'}
  /v1/drivers/me/offer:
    get:
      tags: [Drivers]
      operationId: getCurrentOffer
      summary: The pending offer, if any
      description: V1 drivers poll this; from V2 apps call it on every connect. The first fetch marks the offer as seen.
      responses:
        '200':
          description: Pending offer.
          headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/Offer'}
        '204': {$ref: '#/components/responses/NoContent'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
...
  /v1/realtime/tickets:
    post:
      tags: [Realtime]
      operationId: createRealtimeTicket
      x-since: V2
      summary: (V2) A single-use ticket for the WebSocket handshake
      description: Valid for 60 s. Connect to `url`; message schemas are in `docs/schemas/websocket/`.
      responses:
        '201':
          description: Ticket.
          headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
          content:
            application/json:
              schema: {$ref: '#/components/schemas/RealtimeTicket'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '429': {$ref: '#/components/responses/TooManyRequests'}

  # ------------------------------------------------------------------ Operations
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/v1/auth/token` or `/v1/auth/refresh`.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Unique per logical command; reuse it only to retry the same request.
      schema: {type: string, minLength: 1, maxLength: 255, pattern: '^[\x21-\x7E]+$'}
...
  headers:
    RequestId:
      required: true
      description: Correlation ID of this request.
      schema: {type: string}
    IdempotentReplayed:
      description: '`true` when this is the stored response of an earlier request with the same key.'
      schema: {type: string, enum: ['true', 'false']}
    RetryAfter:
      description: Seconds to wait before retrying.
      schema: {type: integer, minimum: 1}
  responses:
    NoContent:
      description: Nothing to return.
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
...
    DriverStatus:
      description: The driver's availability.
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Idempotent-Replayed: {$ref: '#/components/headers/IdempotentReplayed'}
      content:
        application/json:
          schema: {$ref: '#/components/schemas/DriverStatus'}
...
    BadRequest:
      description: '`VALIDATION_FAILED` (see `errors`), `MALFORMED_REQUEST` or `IDEMPOTENCY_KEY_REQUIRED`.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Unauthorized:
      description: '`UNAUTHENTICATED`, `CODE_INVALID`, `REFRESH_TOKEN_INVALID` or `WEBHOOK_SIGNATURE_INVALID`.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Forbidden:
      description: '`FORBIDDEN` (wrong role) or `ACCOUNT_DISABLED`.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    NotFound:
      description: '`NOT_FOUND`: unknown, or not visible to the caller.'
      headers: {X-Request-Id: {$ref: '#/components/headers/RequestId'}}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    Conflict:
      description: |
        The state doesn't allow this now; re-read the resource. Codes: `INVALID_TRANSITION`, `RIDE_REASSIGNED`,
        `OFFER_NO_LONGER_AVAILABLE`, `QUOTE_EXPIRED`, `QUOTE_ALREADY_USED`, `ACTIVE_RIDE_EXISTS`, `DUES_OUTSTANDING`,
        `PIN_LOCKED`, `NO_SHOW_TOO_EARLY`, `DRIVER_NOT_ELIGIBLE`, `DRIVER_HAS_ACTIVE_RIDE`, `PLACES_LIMIT_REACHED`,
        `NO_DUES`, `CHARGE_NOT_REFUNDABLE`, `RATING_NOT_OPEN`, `ALREADY_RATED`, `RECEIPT_NOT_AVAILABLE`,
        `VERSION_CONFLICT` (with `current_version`), `ALREADY_EXISTS`, `FLAG_ALREADY_RESOLVED`, or
        `IDEMPOTENCY_KEY_IN_PROGRESS` (retry with the same key after `Retry-After`).
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    TooManyRequests:
      description: '`RATE_LIMITED` or `CODE_ATTEMPTS_EXCEEDED`. Not processed, and the `Idempotency-Key` was not consumed.'
      headers:
        X-Request-Id: {$ref: '#/components/headers/RequestId'}
        Retry-After: {$ref: '#/components/headers/RetryAfter'}
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
  schemas:
...
    Uuid: {type: string, format: uuid}
    Timestamp: {type: string, format: date-time}
    CityId: {type: string, pattern: '^[a-z]{3,8}$', examples: [blr]}
    CategoryCode: {type: string, pattern: '^[A-Z][A-Z0-9_]{1,19}$', examples: [MINI]}
...
    Money:
      type: object
      required: [amount_paise, currency]
      properties:
        amount_paise: {type: integer, format: int64, description: 'Paise; negative only in earnings adjustments.'}
        currency: {type: string, pattern: '^[A-Z]{3}$', examples: [INR]}
    GeoPoint:
      type: object
      required: [lat, lon]
      properties:
        lat: {type: number, minimum: -90, maximum: 90}
        lon: {type: number, minimum: -180, maximum: 180}
    RatingSummary:
      type: object
      required: [count]
      properties:
        average: {type: number, minimum: 1, maximum: 5, description: Average of the latest 100 ratings; absent with no ratings.}
        count: {type: integer, minimum: 0, maximum: 100}
    Problem:
      type: object
      required: [title, status, code, request_id]
      properties:
        type: {type: string, default: about:blank}
        title: {type: string}
        status: {type: integer}
        detail: {type: string}
        instance: {type: string}
        code: {type: string, description: Stable machine-readable code (LLD §13.2).}
        request_id: {type: string}
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
        current_status: {type: string}
        current_version: {type: integer}
        attempts_left: {type: integer, minimum: 0}
        available_at: {$ref: '#/components/schemas/Timestamp'}
        dues: {$ref: '#/components/schemas/Money'}
...
    AvailabilityStatus: {type: string, enum: [OFFLINE, AVAILABLE, OFFERED, ASSIGNED, ON_TRIP]}
    OfferStatus: {type: string, enum: [PENDING, ACCEPTED, DECLINED, EXPIRED, WITHDRAWN]}
...
    Verification: {type: string, enum: [PENDING, VERIFIED, REJECTED]}
...
    PersonSummary:
      type: object
      required: [id, first_name]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        first_name: {type: string}
        rating: {$ref: '#/components/schemas/RatingSummary'}
...
    Vehicle:
      type: object
      required: [id, driver_id, category, plate, make, model, colour, active]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        driver_id: {$ref: '#/components/schemas/Uuid'}
        category: {$ref: '#/components/schemas/CategoryCode'}
        plate: {type: string}
        make: {type: string}
        model: {type: string}
        colour: {type: string}
        active: {type: boolean}
        version: {type: integer}
    DriverStatus:
      type: object
      required: [driver_id, status, version]
      properties:
        driver_id: {$ref: '#/components/schemas/Uuid'}
        status: {$ref: '#/components/schemas/AvailabilityStatus'}
        city_id: {$ref: '#/components/schemas/CityId'}
        vehicle_id: {$ref: '#/components/schemas/Uuid'}
        category: {$ref: '#/components/schemas/CategoryCode'}
        online_since: {$ref: '#/components/schemas/Timestamp'}
        offer_id: {$ref: '#/components/schemas/Uuid'}
        ride_id: {$ref: '#/components/schemas/Uuid'}
        version: {type: integer, minimum: 0}
...
    Driver:
      type: object
      required: [id, first_name, city_id, verification, suspended, vehicles, status]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        first_name: {type: string}
        last_name: {type: string}
        city_id: {$ref: '#/components/schemas/CityId'}
        verification: {$ref: '#/components/schemas/Verification'}
        suspended: {type: boolean}
        rating: {$ref: '#/components/schemas/RatingSummary'}
        vehicles: {type: array, items: {$ref: '#/components/schemas/Vehicle'}}
        status: {$ref: '#/components/schemas/DriverStatus'}
    GoOnlineRequest:
      type: object
      required: [vehicle_id]
      properties:
        vehicle_id: {$ref: '#/components/schemas/Uuid'}
    LocationUpdate:
      type: object
      required: [seq, lat, lon, accuracy_m, device_time]
      properties:
        seq: {type: integer, format: int64, minimum: 0, description: Increases with every update the app records.}
        lat: {type: number, minimum: -90, maximum: 90}
        lon: {type: number, minimum: -180, maximum: 180}
        accuracy_m: {type: number, minimum: 0}
        heading_deg: {type: number, minimum: 0, exclusiveMaximum: 360}
        speed_mps: {type: number, minimum: 0}
        device_time: {$ref: '#/components/schemas/Timestamp'}
        replay: {type: boolean, default: false}
    LocationBatch:
      type: object
      required: [updates]
      properties:
        updates: {type: array, minItems: 1, maxItems: 100, items: {$ref: '#/components/schemas/LocationUpdate'}}
    LocationBatchResult:
      type: object
      required: [applied, stale, ignored]
      properties:
        applied: {type: integer, minimum: 0}
        stale: {type: integer, minimum: 0, description: Not newer than the last applied sequence number.}
        ignored: {type: integer, minimum: 0, description: Driver offline, out of bounds or poor quality.}
        last_applied_seq: {type: integer, format: int64}
    Offer:
      type: object
      required: [id, ride_id, status, category, pickup, dropoff, pickup_distance_m, fare, created_at, expires_at]
      properties:
        id: {$ref: '#/components/schemas/Uuid'}
        ride_id: {$ref: '#/components/schemas/Uuid'}
        status: {$ref: '#/components/schemas/OfferStatus'}
        category: {$ref: '#/components/schemas/CategoryCode'}
        pickup: {$ref: '#/components/schemas/GeoPoint'}
        dropoff: {$ref: '#/components/schemas/GeoPoint'}
        pickup_distance_m: {type: integer, minimum: 0}
        fare: {$ref: '#/components/schemas/Money'}
        rider: {$ref: '#/components/schemas/PersonSummary'}
        created_at: {$ref: '#/components/schemas/Timestamp'}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
        expires_in_ms: {type: integer, minimum: 0, description: Computed by the server, so a wrong device clock can't change the countdown.}
...
    RealtimeTicket:
      type: object
      required: [ticket, expires_at, url]
      properties:
        ticket: {type: string}
        expires_at: {$ref: '#/components/schemas/Timestamp'}
        url: {type: string, format: uri}

    # ---------------------------------------------------------------- operations
...